[Ide] Fix New Project dialog's language selector being truncated
The size of the language button cell renderer did not increase its
size for any text it needed to contain.

Still some problems here when compared with version 7.1, before the
language button was moved to its own cell renderer. In 7.1 the
drop down resizes to the left, so the right hand edge is fixed in
position. Also the recently used templates language moves to the left
after selecting another category and going back to the recently used
list. Once the language cell renderer gets an increased width it does
not seem to reduce the width again even though GetSize returns a
smaller width.

Fixes VSTS 738565 - C#/F# chooser in New Project should be moved a few
pixels leftward